refactor: extract report logic
