+++ src/app.tsx
@@ -1,22 +1,7 @@
 import { useState } from "react";
-
-// 現在時刻を30分単位で丸めてHH:MM形式で返す
-const roundedNow = () => {
-  const now = new Date();
-  const minutes = now.getHours() * 60 + now.getMinutes();
-  const rounded = Math.round(minutes / 30) * 30;
-  const h = Math.floor(rounded / 60) % 24;
-  const m = rounded % 60;
-  return `${h.toString().padStart(2, "0")}:${m.toString().padStart(2, "0")}`;
-};
-
-// タスク型定義
-type Task = {
-  id: string;
-  start: string;
-  name: string;
-  end: string;
-};
+import { roundedNow } from "./utils/time";
+import { Task } from "./types";
+import { generateReport } from "./report";
 
 function App() {
   const [tasks, setTasks] = useState<Task[]>(() => {
@@ -57,52 +42,13 @@ function App() {
   };
 
   // 日報出力用データ生成
-  const generateReport = () => {
+  const handleGenerateReport = () => {
     if (tasks.length === 0) return;
     if (hasIncomplete) {
       alert("開始時刻と終了時刻を入力してください");
       return;
     }
-    // 日付・出勤・退勤・休憩・合計稼働・タスク集計
-    const sorted = [...tasks].sort((a, b) => a.start.localeCompare(b.start));
-    const date = new Date();
-    const y = date.getFullYear();
-    const m = date.getMonth() + 1;
-    const d = date.getDate();
-    const workStart = sorted[0].start;
-    const workEnd = sorted[sorted.length - 1].end;
-    // 休憩時間合計
-    let breakMinutes = 0;
-    const workMap: Record<string, number> = {};
-    sorted.forEach(t => {
-      const start = t.start;
-      const end = t.end;
-      if (!start || !end) return;
-      const [sh, sm] = start.split(":").map(Number);
-      const [eh, em] = end.split(":").map(Number);
-      const min = (eh * 60 + em) - (sh * 60 + sm);
-      if (t.name === "休憩") {
-        breakMinutes += min;
-      } else {
-        workMap[t.name] = (workMap[t.name] || 0) + min;
-      }
-    });
-    // 合計稼働時間
-    const [sh, sm] = workStart.split(":").map(Number);
-    const [eh, em] = workEnd.split(":").map(Number);
-    const totalMinutes = (eh * 60 + em) - (sh * 60 + sm) - breakMinutes;
-    // フォーマット関数
-    const fmt = (min: number) => `${Math.floor(min/60)}:${(min%60).toString().padStart(2, "0")}`;
-    // 出力
-    let txt = `日付: ${y}/${m}/${d}\n`;
-    txt += `出勤時刻: ${workStart}\n`;
-    txt += `退勤時刻: ${workEnd}\n`;
-    txt += `休憩時間: ${fmt(breakMinutes)}\n`;
-    txt += `合計稼働時間: ${fmt(totalMinutes)}\n`;
-    txt += `タスク一覧\n`;
-    Object.entries(workMap).forEach(([name, min]) => {
-      txt += `  - 名前: ${name} / 総作業時間: ${fmt(min as number)}\n`;
-    });
+    const txt = generateReport(tasks);
     setResult(txt);
   };
 
@@ -152,7 +98,7 @@ function App() {
         <button onClick={addTask} style={{ marginTop: 8 }}>＋タスク追加</button>
       </div>
       <div style={{ margin: "1.5rem 0" }}>
-        <button onClick={generateReport} disabled={isMultiDay || hasIncomplete || tasks.length === 0} style={{ fontSize: "1.1em", padding: "0.5em 1.5em" }}>
+          <button onClick={handleGenerateReport} disabled={isMultiDay || hasIncomplete || tasks.length === 0} style={{ fontSize: "1.1em", padding: "0.5em 1.5em" }}>
           保存
         </button>
         {isMultiDay && <span style={{ color: "red", marginLeft: 12 }}>日付をまたぐタスクが含まれています</span>}
