+++ src/types.ts
@@ -1,0 +1,7 @@
+export type Task = {
+  id: string;
+  start: string;
+  name: string;
+  end: string;
+};
+
